==> Scripts/0.AdaptationElementsLinkages_Processing.Rmd <==
---
title: "AdaptationElementsLinkages_Processing"
output: html_document
date: "2023-10-23"
---
# Summary
Data for this code are available at: https://doi.org/10.7910/DVN/VK3CP, file "AdaptationElementsLinkages".
This code generates an adjacency matrix based on information on linkages between the adaptation elements.  
The resulting dataframe should be saved to a local file, as it will be used as basis for creating the chord diagram (See code "2.Coherence")

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load package
```{r}
library(tidyr)
library(dplyr)
```
# Preparation
To run this code, download the file "AdaptationElementsLinkages" available at: https://doi.org/10.7910/DVN/VK3CP.
Adjust the file path below with the correct location where you saved the "AdaptationElementsLinkages" file
```{r}
linkages <- read.csv(file = "C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Data/Data-Raw/AdaptationElementsLinkages.csv", fileEncoding = "Windows-1252")
```

# Step 1: Dataset Examination
We began with the "AdaptationElementsLinkages" dataset containing information on the connections between different elements such as 'Hazard', 'System', 'Goal', etc., for 53 countries. Each row represented a country's document (NAP and/or NDC), and the columns indicated binary relations (1 for a connection and 0 for no connection) between these elements.

# Step 2: Aggregation
Given that a country may have multiple entries (e.g., both NDC and NAP), we first aggregate the data by country. In this aggregation process, we don't just simply sum the occurrences but rather mark the presence (1) if any of the country's documents mentioned the relationship and absence (0) if neither did. This is done to avoid double counting and to ensure that the data accurately represents whether a country has considered a specific relationship in either of its documents.
```{r}
linkages_agg <- linkages %>%
   group_by(Country, ISO3) %>%
  summarise(across(HazardSystem:ActionIndicator, ~ifelse(any(. == 1), 1, 0)))
```

# Step 3: Construction of Adjacency Matrix
An empty 6x6 adjacency matrix is initialized to capture the relationships between pairs of elements. This matrix's rows and columns represents the distinct elements ('Hazard', 'System', 'Goal', etc.), and the matrix would eventually show how often each pair of elements co-occurred in the dataset.
```{r}
# Create an empty 6x6 matrix
nodes <- c("Hazard", "System", "Goal", "Objective", "Action", "Indicator")
adj_matrix <- matrix(0, nrow = 6, ncol = 6, dimnames = list(nodes, nodes))

# Extract the connection columns
connection_columns <- colnames(linkages_agg)[5:ncol(linkages_agg)]

# Define a function to map column names to nodes
map_columns_to_nodes <- function(column_name) {
  unlist(strsplit(column_name, split = "(?<=\\w)(?=\\p{Lu})", perl = TRUE))
}
```

# Step 4: Matrix Population
We loop through the columns of the "linkages-agg" dataset to extract relationship information.
For each column, the element pair it represents is identified (e.g., the column "HazardSystem" represents a connection between 'Hazard' and 'System').
The values in each column are summed up, providing a count of occurrences for that specific relationship across all documents.
The resulting sum is then added to the corresponding cell in the adjacency matrix.
Adjustments: We discovered discrepancies for specific element pairs in the matrix. To correct these, we directly computed the sums for the problematic relationships from the  dataframe and manually set the correct values in the adjacency matrix.
The final adjacency matrix provides a clear, summarized view of how often each pair of elements was connected across all documents in the dataset.
```{r}
# Fill in the matrix using the data
for (colname in connection_columns) {
  nodes_involved <- map_columns_to_nodes(colname)
  connection_value <- sum(linkages_agg[[colname]], na.rm = TRUE)
  
  if (!identical(nodes_involved[1], nodes_involved[2])) {
    adj_matrix[nodes_involved[1], nodes_involved[2]] <- connection_value
    adj_matrix[nodes_involved[2], nodes_involved[1]] <- connection_value
  }
}
# Adjusting values for problematic entries
adj_matrix["Hazard", "System"] <- sum(linkages_agg$HazardSystem, na.rm = TRUE)
adj_matrix["System", "Hazard"] <- sum(linkages_agg$HazardSystem, na.rm = TRUE)

adj_matrix["Hazard", "Goal"] <- sum(linkages_agg$HazardGoal, na.rm = TRUE)
adj_matrix["Goal", "Hazard"] <- sum(linkages_agg$HazardGoal, na.rm = TRUE)

# Print the adjacency matrix
print(adj_matrix)
```

# Descriptive statistics
Compute information on number of linkages per country
```{r}
linkages_agg_total <- 
  linkages_agg %>%
  ungroup() %>%
  mutate(count_1s = rowSums(select(., HazardSystem:ActionIndicator) == 1))
```

# Save data to a local file
Modify the file path to specify the location where you would like to store the new dataframe
```{r}
write.csv(adj_matrix,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Data/Data-Processed/LinkagesMatrix.csv")
write.csv(linkages_agg_total, "C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_LinkagesXcountry.csv")
```
