==> Scripts/1.Comprehensiveness-Elements.Rmd <==
---
title: "Comprehensiveness-Elements"
output: html_document
date: "2023-02-03"
---

# Summary
Data for this code are available at: https://doi.org/10.7910/DVN/VK3CP
This code generates assesses comprehensiveness of information across countries and documents (NDCs and NAPs). 
Outputs include descriptive statistics on counts and distributions of adaptation tracking elements across documents, countries, and sectors, as well as figures and tables reported in the manuscript file, extended data file and supplementary information file.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load packages
```{r}
library(tidyverse)      
library(forcats)      
library(dplyr)
library(ggplot2)
library(tidyr)
library(geomtextpath)
library(ggrepel)
```

# Preparation
To run this code, download the file "AdaptationElements" available at: https://doi.org/10.7910/DVN/VK3CP.
Adjust the file path below with the correct location where you saved the "AdaptationElements" file
```{r}
adaptdata <- read.csv(file = "C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Data/Data-Processed/AdaptationElementsProcessed.csv", fileEncoding = "Windows-1252")
```

# Descriptive statistics
## n elements: all countries, all documents
```{r}
all_elem <- adaptdata[, c("Document", "Element")]
all_elem$Total <- 1

# Summary of elements
nElements <- all_elem %>%
  group_by(Element) %>%
  summarise(
    Total = sum(Total))

# how many countries with NAPs
number_of_countries_with_NAPs <- adaptdata %>%
   filter(Document == "NAP") %>%
  distinct(Country) %>%
  pull(Country)
print(number_of_countries_with_NAPs)
```

## n elements: all countries, by document
```{r}
# Summary of elements by document
nElements_doc <- all_elem %>%
  group_by(Element, Document) %>%
  summarise(
    Total = sum(Total))
```

## n elements: by country
```{r}
# add cols for  elements, convert to numeric
adaptdata <- adaptdata %>%
  mutate(
    Hazards = as.numeric(Element == "Hazard"),
    SystemsAtRisk = as.numeric(Element == "System at risk"),
    Goals = as.numeric(Element == "Goal"),
    Objectives = as.numeric(Element == "Objective"),
    Actions = as.numeric(Element == "Action"),
    Indicators = as.numeric(Element == "Indicator")
  )

#select columns of interest
summary_elem <- adaptdata[, c("Country", "Document", "Hazards", "SystemsAtRisk", "Goals", "Objectives", "Actions", "Indicators")]

# create a table to summarize counts of elements by country
nelem_cty <- summary_elem %>%
  group_by(Country) %>%
  summarise(
    Hazards = sum(Hazards),
    Systems = sum(SystemsAtRisk),
    Goals = sum(Goals),
    Objectives = sum(Objectives),
    Actions = sum(Actions),
    Indicators = sum(Indicators))

nelem_cty$Country[grepl("Ivoire", nelem_cty$Country)] <- "Cote d'Ivoire"
# How many countries have all elements
cty_all_elem <- nelem_cty %>% filter_all(all_vars(. > 0))
```

## n elements and n countries: by element type 
```{r}
# how many countries mention an element type
n_cty_element <- adaptdata  %>%
  group_by(Element) %>%
  summarise(nCountries = n_distinct(Country)) %>%
  arrange(-nCountries)

# how many observations per element
n_elem_type <- adaptdata %>%
  count(Element, name="Observations")

# merge datasets
nElem_nCty <- merge(n_cty_element, n_elem_type, all = TRUE)
```

## summary stats: n element types and n element counts
```{r}
# Summary statistics: element types in a country
presence_elem_cty <- nelem_cty%>% mutate_at(vars(Hazards, Systems, Goals, Objectives, Actions, Indicators), 
            ~ ifelse(. >= 1, 1, 0))

## calculate sum per row (per country)
presence_elem_cty <- presence_elem_cty %>%
  mutate(Total_Mentions = Hazards + Systems + Goals + Objectives + Actions + Indicators)  

stats_elem_presence <- presence_elem_cty %>%
  summarise(
    Average = mean(Total_Mentions),
    Median = median(Total_Mentions),
    Minimum = min(Total_Mentions),
    Maximum = max(Total_Mentions)
  )

# Summary statistics: elements counts
nelem_cty <- nelem_cty %>%
  mutate(Total_Elements = Hazards + Systems + Goals + Objectives + Actions + Indicators)  

stats_elem_count_cty <- nelem_cty %>%
  summarise(
    Average = mean(Total_Elements),
    Median = median(Total_Elements),
    Minimum = min(Total_Elements),
    Maximum = max(Total_Elements)
  )
```

## n elements: by sector
```{r}
#subset columns of interest
sector_subset <- adaptdata[, c("Country", "Document", "SectorType", "Hazards", "SystemsAtRisk", "Goals", "Objectives", "Actions", "Indicators")]
sector_subset$Country[sector_subset$Country == "C<f4>te d'Ivoire"] <- "Cote d'Ivoire"

#summarize by sector
nelem_sec <- sector_subset %>%
  group_by(SectorType) %>%
  summarise(
    Hazards = sum(Hazards),
    Systems = sum(SystemsAtRisk),
    Goals = sum(Goals),
    Objectives = sum(Objectives),
    Actions = sum(Actions),
    Indicators = sum(Indicators))

nelem_sec <- nelem_sec %>%
  mutate(Total_Elements = Hazards + Systems + Goals + Objectives + Actions + Indicators)  

summaries <- apply(nelem_sec[, 2:8], 2, sum)
new_row <- c("All_Sectors", summaries)
nelem_sec <- rbind (nelem_sec, new_row)
```

## n elements: by country and sector 
```{r}
nelem_cty_sec <- sector_subset %>%
  group_by(Country, SectorType) %>%
  summarise(Total_Elements = sum(Hazards, na.rm = TRUE) +
                              sum(SystemsAtRisk, na.rm = TRUE) +
                              sum(Goals, na.rm = TRUE) +
                              sum(Objectives, na.rm = TRUE) +
                              sum(Actions, na.rm = TRUE) +
                              sum(Indicators, na.rm = TRUE)) %>%
  pivot_wider(names_from = SectorType, values_from = Total_Elements) %>%
  mutate_all(~replace_na(., 0)) 
```

## n countries: per element and sector 
```{r}
group_sector_cty_elem <- sector_subset %>%
  group_by(SectorType, Country) %>%
  summarise(
    Hazards = sum(Hazards),
    Systems = sum(SystemsAtRisk),
    Goals = sum(Goals),
    Objectives = sum(Objectives),
    Actions = sum(Actions),
    Indicators = sum(Indicators))

# How many countries have an element: Summarize by sector
ncty_sec_elem <- group_sector_cty_elem %>%
  group_by(SectorType) %>%
  summarise(
   Hazards = sum(Hazards > 0),
    Systems = sum(Systems > 0),
    Goals = sum(Goals > 0),
    Objectives = sum(Objectives > 0),
    Actions = sum(Actions > 0),
    Indicators = sum(Indicators > 0))

# How many countries have an element: Summarize for all sectors
ncty_elem_all <- nelem_cty %>%
  summarise(
    SectorType = "All_Sectors",
    Hazards = sum(Hazards > 0),
    Systems = sum(Systems > 0),
    Goals = sum(Goals > 0),
    Objectives = sum(Objectives > 0),
    Actions = sum(Actions > 0),
    Indicators = sum(Indicators > 0)
  )

# Append to the sector summary
ncty_sec_elem <- bind_rows(ncty_sec_elem, ncty_elem_all)
```
## n countries per sector
```{r}
# How many countries for each sector
ncty_sec <- group_sector_cty_elem %>%
  group_by(SectorType) %>%
  summarise(unique_countries = n_distinct(Country)) %>%
  arrange(-unique_countries) 
```

## n sectors per country
```{r}
# only take into account sectors that are being explicitly defined
nsector_cty <- sector_subset %>% filter(SectorType != "Not specified")

nsector_cty <- nsector_cty %>%
  group_by(Country) %>%
  summarise(unique_sectors = n_distinct(SectorType)) %>%
  arrange(-unique_sectors) 

stats_sec_cty <- nsector_cty %>%
  summarise(
    Average = mean(unique_sectors),
    Median = median(unique_sectors),
    Minimum = min(unique_sectors),
    Maximum = max(unique_sectors))
```

## agriculture: n observations and n countries per element
```{r}
# subset data for agriculture (food, fiber and other ecosystem products)
agadapt <- adaptdata[adaptdata$SectorType %in% c("Food, fiber, other ecosystem products"), ]

# how many countries mention an element type
n_cty_element_ag <- agadapt  %>%
  group_by(Element) %>%
  summarise(nCountries = n_distinct(Country)) %>%
  arrange(-nCountries)

# how many observations per element
n_elem_type_ag <- agadapt %>%
  count(Element, name="Elements")

# merge datasets
merged_df_ag <- merge(n_cty_element_ag, n_elem_type_ag, all = TRUE)
```

### Save data to local files
Modify the file path to specify the location where you would like to store the new dataframe
```{r}
write.csv(nElements,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElements_aggregate.csv")
write.csv(nElements_doc,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElementsXdocument.csv")
write.csv(nelem_cty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElementsXcountry.csv")
write.csv(cty_all_elem,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nCountriesAllElements.csv")
write.csv(nElem_nCty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElementsXnCountry.csv")
write.csv(merged_df_ag,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElementsXnCountry_Ag.csv")
write.csv(stats_elem_presence,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_ElemTypesMinMax.csv")
write.csv(stats_elem_count_cty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_ElemCountsMinMax.csv")
write.csv(nelem_sec,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElementsXsector.csv")
write.csv(nelem_cty_sec,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nElementsXsectorXcty.csv")
write.csv(ncty_sec_elem,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nCountriesXsectorXelement.csv")
write.csv(ncty_sec,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nCountriesXsector.csv")
```

# Manuscript figures
Code to generate figures and tables in the main file, extended data file and supplementary information files
## Figure 1. Heatmap with countries and element Counts
```{r}
# adjust names of the elements to match the number of total elements
long_cty_elem <- nelem_cty %>%
  gather(key = "Type", value = "Value", -Country)

# adjust names of countries so that they fit on the plot
long_cty_elem$Country[grepl("Democratic Republic of the Congo", long_cty_elem$Country)] <- "Dem. Rep. of the Congo"

# remove rows with "Total elements"
long_cty_elem <- long_cty_elem[-grep("Total_Elements", long_cty_elem$Type), ]

# add elements
# Hazards (357), Systems at risk (478), Goals (64),Objectives (1271), Actions (4441), Indicators (357)

# Set factor levels for Category
long_cty_elem$Type <- factor(long_cty_elem$Type, levels = c("Hazards", "Systems", "Goals", "Objectives", "Actions", "Indicators"))
```
prepare plot
```{r}
# Define break points and color labels
break_points <- c(-Inf, 0, 50, 100, 150, 200, 250, 300, 350)
color_labels <- c("0", "1-50", "51-100", "101-150", "151-200", "201-250", "251-300", "301-350")

# Generate the color_class column
long_cty_elem$color_class <- cut(long_cty_elem$Value, 
                                 breaks = break_points, 
                                 labels = color_labels, 
                                 right = TRUE, include.lowest = TRUE)

# Compute mid-points for color_labels
mid_points <- sapply(2:length(break_points), function(i) {
  (break_points[i-1] + break_points[i]) / 2
})

# Adjusting for the -Inf value in the first break point
mid_points[1] <- 0

# Create the color_value column
long_cty_elem$color_value <- as.numeric(mid_points[as.numeric(long_cty_elem$color_class)])

# Define the color palette
greens <- c("#D0E9E7","#8bcac4", "#5ab4ac", "#01665e", "#004741","#00332f", "#001a18" )
colors <- c("white", greens)
```
plot
```{r}
fig1 <- ggplot(long_cty_elem, aes(x = Type, y = Country, fill = color_value)) + 
  geom_tile(aes(width = 1, height = 1), color = "grey85", size = 0.1) + # insert white line between the cells
  scale_fill_gradientn(colors = colors, 
                       breaks = mid_points,
                       labels = color_labels,
                       na.value = "white") +
  scale_y_discrete(limits = rev(unique(long_cty_elem$Country))) + 
  theme_bw() +
  guides(fill = guide_legend(direction = "horizontal", title.position = "top", keywidth = unit(1, "cm"), keyheight = unit(0.3, 'cm'))) +
  theme(
    panel.border = element_blank(),
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),
    axis.text.x.top = element_blank(),
    axis.title.x = element_blank(),
    axis.title.y = element_blank(),
    axis.ticks.x.top = element_blank(), 
    axis.ticks.x.bottom = element_blank(), 
    legend.position = "bottom",
    legend.box.background = element_rect(color = "white", size = 0.5),
    legend.box.margin = margin(0, 0, 0, 0),
    legend.key.width = unit(1, 'cm'),
    legend.key.height = unit(0.3, 'cm'),
    legend.spacing.x = unit(0.2, 'cm'),
    strip.background = element_rect(fill = "white", color = "white", size = 0.1), 
    strip.text = element_blank()
  ) +
  labs(fill = "") +
  scale_x_discrete(position = "top")
fig1

# Save figure to local file
# Modify the file path to specify the location where you would like to store the figure
ggsave("C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/Fig1.png", fig1, width=5, height=9)
```

## Figure 2: Elements by type and sector (percentages)
This code gives the table to build Figure 2. 
```{r}
merged_table <- merge(ncty_sec_elem, nelem_sec[, c("SectorType", "Total_Elements")], by = "SectorType", all.x = TRUE)
merged_table$Total_Elements <- as.numeric(merged_table$Total_Elements)
percentage_df <- merged_table
percentage_df[, 2:7] <- round(percentage_df[, 2:7] / 53, 2)  # Divide each value in columns 2 to 7 by 53
percentage_df$Total_Elements <- round(percentage_df$Total_Elements / 7686, 2)  # Divide Total_Elements by 7686

# Save table to local file
# Modify the file path to specify the location where you would like to store the table
write.csv(percentage_df,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/Fig2.csv")
```

## Extended Data Table 2: Distribution of elements (percentages) by sector and country
```{r}
nelem_cty_sec2 <- nelem_cty_sec
colnames(nelem_cty_sec2) <- gsub("[, ]", "_", colnames(nelem_cty_sec2))

nelem_cty_sec2 <- nelem_cty_sec2 %>%
  mutate(Total = Cities__settlements__key_infrastructure + 
           Crosscutting + 
           Livelihoods__poverty__sustainable_development  + 
           Not_specified + 
           Terrestrial__freshwater_ecosystems + 
           Food__fiber__other_ecosystem_products + 
           Health__wellbeing__communities + 
           Ocean__coastal_ecosystems + 
           Water__sanitation__hygiene 
         )  

nelem_cty_sec2 <- nelem_cty_sec2 %>%
  mutate(Pc_Cities_Settlements = round((Cities__settlements__key_infrastructure / Total) * 100, 1),
         Pc_Crosscutting = round((Crosscutting / Total) * 100, 1),
         Pc_Livelihoods_Poverty = round((Livelihoods__poverty__sustainable_development / Total) * 100, 1),
         Pc_Not_specified = round((Not_specified / Total) * 100, 2),
         Pc_Terrestrial_Freshwater = round((Terrestrial__freshwater_ecosystems / Total) * 100, 1),
         Pc_Food_fiber = round((Food__fiber__other_ecosystem_products / Total) * 100, 1),
         Pc_Health_Wellbeing = round((Health__wellbeing__communities / Total) * 100, 1),
         Pc_Ocean_Coastal = round((Ocean__coastal_ecosystems / Total) * 100, 1),
         Pc_Water_Sanitation = round((Water__sanitation__hygiene / Total) * 100, 1))

nelem_cty_sec2 <- nelem_cty_sec2[, c(1, 11:20)]

nelem_cty_sec2 <- nelem_cty_sec2 %>%
  arrange(desc(Total))

# Save table to local file
# Modify the file path to specify the location where you would like to store the table
write.csv(nelem_cty_sec2,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/ED_Table2.csv")
```

## Extended Data Figure 2: pc countries with info on hazards and systems at risk
### 2a. Hazards
```{r}
hazards <- adaptdata[adaptdata$Element == "Hazard", ]

# how many unique countries mention a hazard type
n_cty_hazard <- hazards  %>%
  group_by(HazardType) %>%
  summarise(nCountries = n_distinct(Country)) %>%
  arrange(-nCountries)

# mentions of a hazard by country. 
## contingency table (a cross table) where rows represent countries, columns represent hazard types, and  entries in the table indicate if the hazard type is mentioned (1) or not (0) for each country.
hazard_cty <- hazards  %>%
 select(Country, HazardType) %>%            # Select relevant columns
  distinct() %>%                             # Remove duplicates
  mutate(Mentioned = 1) %>%                  # Assign value of 1 to each occurrence
  spread(key = HazardType, value = Mentioned, fill = 0) # Reshape

total_row_h <- data.frame(
  Country = "Total",
  t(colSums(hazard_cty[, 2:15])))
names(total_row_h) <- gsub("\\.", " ", names(total_row_h))
hazard_cty <- bind_rows(hazard_cty, total_row_h)

# Save tables to local file
# Modify the file paths to specify the location where you would like to store the tables
write.csv(n_cty_hazard,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/ED_nCountriesXhazard.csv")
write.csv(hazard_cty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/ED_HazardXcountries.csv")
```

### 2b. Systems at Risk
```{r}
# subset Systems dataset
systems <- adaptdata[adaptdata$Element == "System at risk", ]

# how many unique countries mention a system type
n_cty_system <- systems  %>%
  group_by(SystemType) %>%
  summarise(nCountries = n_distinct(Country)) %>%
  arrange(-nCountries)

# mentions of a system by country. 
## contingency table (a cross table) where rows represent countries, columns represent system types, and  entries in the table indicate if the system type is mentioned (1) or not (0) for each country.
systems_cty <- systems  %>%
 select(Country, SystemType) %>%            # Select relevant columns
  distinct() %>%                             # Remove duplicates
  mutate(Mentioned = 1) %>%                  # Assign value of 1 to each occurrence
  spread(key = SystemType, value = Mentioned, fill = 0) # Reshape

total_row_s <- data.frame(
  Country = "Total",
  t(colSums(systems_cty[, 2:15])))
names(total_row_s) <- gsub("\\.", " ", names(total_row_s))
systems_cty <- bind_rows(systems_cty, total_row_s)

# Save tables to local file
# Modify the file paths to specify the location where you would like to store the tables
write.csv(n_cty_system,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/ED_nCountriesXsystem.csv")
write.csv(systems_cty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/ED_SystemsXcountries.csv")
```

# Other analyses
## Goal types
```{r}
Goals <- adaptdata[adaptdata$Element == "Goal", ]
Goals <- Goals[, c("Country", "Document", "Element", "ElementText",  "SectorType", "TargetTime")]

Goals$Resilience <- ""
Goals$AdaptiveCap <- ""
Goals$Vulnerability <- ""
  
Goals$Resilience[grep("resilien", Goals$ElementText, ignore.case = TRUE)]  <- 1
Goals$AdaptiveCap[grep("(adapt|capacities|capacity)", Goals$ElementText, ignore.case = TRUE)]  <- 1
Goals$Vulnerability[grep("vulnerab", Goals$ElementText, ignore.case = TRUE)]  <- 1

Goals <- Goals %>%
  mutate_at(vars(Resilience, AdaptiveCap, Vulnerability), as.numeric)
Goals_short <- Goals[, c("Country", "Document", "Resilience", "AdaptiveCap", "Vulnerability")]

goals_cty <- Goals_short %>%
  group_by(Country)  %>%
  summarise(
    nResilience = ifelse(any(Resilience > 0, na.rm = TRUE), 1, 0),
    nAdaptiveCap = ifelse(any(AdaptiveCap > 0, na.rm = TRUE), 1, 0),
    nVulnerability = ifelse(any(Vulnerability > 0, na.rm = TRUE), 1, 0))

total_row_g <- data.frame(
  Country = "Total",
  t(colSums(goals_cty[, 2:4], na.rm = TRUE)))

goals_cty <- bind_rows(goals_cty, total_row_g)

# other goals
Goals_other <- Goals[apply(Goals[, 7:9], 1, function(x) all(is.na(x) | x < 1)), ]

# Save tables to local file
# Modify the file paths to specify the location where you would like to store the tables

write.csv(goals_cty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/Other_GoalTypeCty.csv")
write.csv(Goals_other,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/Other_OtherGoalXcty.csv")
```
