==> Scripts/3.Robustness-Targets-Indicators.Rmd <==
---
title: "3. Robustness_Targets_Indicators"
output: html_document
date: "2023-04-18"
---

# Summary
Data for this code are available at: https://doi.org/10.7910/DVN/VK3CP/, file "AdaptationELements"
This code explains the analysis carried out for assessing robustness of information in NDCs and NAPs. 
Specifically, it was used to generate tables and figures for manuscript files, extended data and supplementry files.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Load packages. Make sure you have them installed.
```{r}
library(dplyr)
library(tidyr)
library(countrycode)
```

# Preparation
To run this code, download the file "AdaptationElements" available at: https://doi.org/10.7910/DVN/VK3CP.
Adjust the file path below with the correct location where you saved the "AdaptationElements" file.
```{r}
adaptdata <- read.csv(file = "C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Data/Data-Processed/AdaptationElementsProcessed.csv")
```

# Targets
## n targets: by type
```{r}
targets <- adaptdata[!is.na(adaptdata$TargetType) & adaptdata$TargetType != "", ]

targets <- targets[, -c(9:10, 19:24)]

# how many targets of each type
nTargets <- targets %>% count(TargetType, sort=TRUE)

# how many targets have a time-frame
nTargetsTime <- targets[targets$TargetType %in% c("measurable and timebound", "timebound"), ]

# how many targets have targets timeframe, by country
n_cty_time <- nTargetsTime %>%
  group_by(Country) %>%
  summarise(nTargetsTime = n(), .groups = 'drop')
```
## n targets: by country and sector
```{r}
tgt_sector <- targets %>%
  group_by(Country, SectorType) %>%
  summarise(count = n()) %>%
  spread(key = SectorType, value = count, fill = 0)
```
## n countries with targets: by target type
```{r}
# how many countries have targets (any type)
n_cty <- targets %>%
  summarise(num_countries = n_distinct(Country)) %>%
  pull(num_countries)
print(n_cty)

# how many unique countries with each target type
n_cty_target <- targets %>%
  group_by(TargetType) %>%
  summarise(nCountries = n_distinct((Country))) %>%
  arrange(-nCountries)
# table with target types and countries
nTargets_nCty <- merge(nTargets, n_cty_target, by = "TargetType")
```
## n countries with targets: by sector
```{r}
n_cty_sector <- targets %>% 
  group_by(SectorType) %>%
  summarise(nCountries = n_distinct(Country), .groups = 'drop')
```

## n targets and countries with targets: by sector 
```{r}
sector_summary <- targets %>%
  group_by(SectorType) %>%
  summarise(nTargets = n(), 
            nCountries = n_distinct(Country), .groups = 'drop')

total_targets <- 2488
total_countries <- 46

sector_summary <- sector_summary %>%
  mutate(pcTargets = round((nTargets / total_targets) * 100),
         pcCountries = round((nCountries / total_countries) * 100))

sector_summary <- sector_summary[, c(1, 2, 4, 3, 5)]
```

## n targets and countries, by timeframe (short, medium, long)
```{r}
# Targets with long, short, medium timeframes
unique_classifications <- unique(nTargetsTime$TargetTime)
print(unique_classifications)

nTargetsSLM <- nTargetsTime %>%
  mutate(TimeClassification = case_when(
    # Long Term
    grepl("long-term|over 10 years|over15years|2040|2050|2040-2050", TargetTime) ~ "long",
    # Medium Term
    grepl("medium-term|2026|2027|2028|2029|2030|2035|5-10 years|10-15years|4-10 years|5-10years|2025-2030|2023-2030|2020-2025, 2025-2030, 2030 -2040|2020-2025, 2025-2030", TargetTime) ~ "medium",
    # Short Term
    grepl("short-term|2020|2021|2022|2023|2024|2025|1-5 years|2-5 years|3-5 years|1-4 years|0-5 years|0-3 years|1-3 years|2-3 years|1-2 years|0-2 years|0-4 years|0-6 years|2-4 years|2-4years|4-5 years|1 year|2 years|3 years|4 years|0-7 years|4-7 years|0-10 years", TargetTime) ~ "short",
    # Handle specific cases that span short to medium terms
    grepl("2023-2024|2022-2030|2022-2025|2021-2026|2021-2030|2019-2021|2020-2022|2020-2030|2020-2025|2015-2030|2021-2024|2021-2027|2021-2022|2023-2027|2024-2026|2021-2025|2021-2023|1-6 years", TargetTime) ~ {
      ifelse(grepl("2026|2027|2028|2029|2030|5-10 years|10-15years|5-10years", TargetTime), "medium", "short")
    },
    TRUE ~ NA_character_  # For any unclassified rows, but you may want to inspect these
  ))

# how many targets of each type and how many countries
nTargetTime_cty <- nTargetsSLM %>%
  group_by(TimeClassification) %>%
  summarise(
    number_of_targets = n(),
    number_of_countries = n_distinct(Country)
  ) %>%
  filter(!is.na(TimeClassification))
```

## agriculture: n targets
```{r}
# targets for ag sector
targets_ag <- targets[targets$SectorType %in% c("Food, fiber, other ecosystem products"),]
# how many targets of each type
nTargets_ag <- targets_ag %>% count(TargetType, sort=TRUE)
```

### Save data to local files
Modify the file path to specify the location where you would like to store the new dataframe
```{r}
write.csv(tgt_sector,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nTargetsXsectorXcountry.csv")
write.csv(nTargets_nCty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nTargetsXnCountries.csv")
write.csv(n_cty_sector,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nCountriesTargetsXsector.csv")
write.csv(sector_summary,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nTargetsCountriesXSector.csv")
write.csv(targets_ag,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_TargetsAg.csv")
```

## Supplementary Table S2: targets for food and fiber, measurable and timebound
```{r}
targets_mt <- targets[targets$TargetType == "measurable and timebound" & targets$SectorType == "Food, fiber, other ecosystem products", ]
targets_mt$ISO3 <- countrycode(targets_mt$Country, "country.name", "iso3c")

targets_mt <- targets_mt[, c(2:5, 7, 9: 13, 16, 17)]
targets_mt_reorder <- targets_mt[, c(12, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11)]

# Save table to local file
# Modify the file path to specify the location where you would like to store the table
write.csv(targets_mt_reorder,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/SI_Table2.csv")
```

# Indicators
## n indicators: by sector
```{r}
indicators <- adaptdata[adaptdata$Element == "Indicator", ]

ind_sector <- indicators %>%
  group_by(SectorType) %>%
  summarise(nIndicators = n(), 
            nCountries = n_distinct(Country), .groups = 'drop')
```

## n indicators: by M&E function
```{r}
ind_cty_type <- indicators %>%
  group_by(IndicatorType) %>%
  summarise(nIndicators = n(), 
            nCountries = n_distinct(Country), .groups = 'drop')

total_indicators <- 1075
total_countries <- 18 

ind_cty_type <- ind_cty_type %>%
  mutate(pcIndicators = round((nIndicators / total_indicators) * 100),
         pcCountries = round((nCountries / total_countries) * 100))

ind_cty_type <- ind_cty_type[, c(1, 2, 4, 3, 5)]
```

## n indicators by M&E function and document
```{r}
ind_funct <- indicators[, c("Document", "IndicatorType")]

ind_funct_doc <- ind_funct %>%
  group_by(Document, IndicatorType) %>%
  summarise(Count = n()) %>%
  pivot_wider(names_from = Document, values_from = Count, values_fill = 0) %>%
  mutate(Total = NDC + NAP,
         Percentage_NDC = round((NDC / Total) * 100),
         Percentage_NAP = round((NAP / Total) * 100))
```

## n indicators by SMART criteria
```{r}
columns_to_replace <- c("Specific", "Measurable", "Achievable", "Relevant", "Timebound")

# replace yes and no with 1 and 0
for (col in columns_to_replace) {
  indicators[[col]] <- as.numeric(ifelse(indicators[[col]] == "yes", 1, 0))
}

# indicators and nr of criteria met
indicators$nr_criteria <- rowSums(indicators[, 20:24])

indicators_criteria_nr <- indicators %>%
  group_by(nr_criteria) %>%
  summarize(nr_indicators = n())

# summary statistics
summary_stats_criteria <- indicators %>%
  summarise(
    Average = mean(nr_criteria),
    Median = median(nr_criteria),
    Minimum = min(nr_criteria),
    Maximum = max(nr_criteria)
  )

# Indicators that meet all SMART criteria
indicators$all_criteria <- ifelse(indicators$Specific == 1 & 
                                  indicators$Measurable == 1 &  
                                  indicators$Achievable == 1 & 
                                  indicators$Relevant == 1 & 
                                  indicators$Timebound == 1, 1, 0)

indicators_smart <- indicators[indicators$SMART == 1, ]

# Number of indicators by smart criteria and country, convert from wide to long format
indicators_subset <- indicators[, c(2, 20:24)]
long_ind <- indicators_subset %>%
  pivot_longer(cols = -Country, names_to = "criteria", values_to = "value")

ind_smart_cty <- long_ind %>%
  group_by(criteria) %>%
  summarise(
    nIndicators = sum(value),
    nCountries = n_distinct(Country[value == 1])
  )

total_indicators <- 1075
total_countries <- 18 

ind_smart_cty <- ind_smart_cty %>%
  mutate(pcIndicators = round((nIndicators / total_indicators) * 100),
         pcCountries = round((nCountries / total_countries) * 100))

ind_smart_cty <- ind_smart_cty[, c(1, 2, 4, 3, 5)]
```

## n indicators: by SMART criteria and document
```{r}
ind_smart_new <- indicators[, c(3, 20:24)]

# Reshape, count, and pivot
ind_smart_doc <- ind_smart_new %>%
  pivot_longer(cols = c(Specific, Measurable, Achievable, Relevant, Timebound), 
               names_to = "Criteria", values_to = "Value") %>%
  group_by(Criteria, Document) %>%
  summarise(Total = sum(Value == 1)) %>%
  pivot_wider(names_from = Document, values_from = Total, values_fill = 0) %>%
  mutate(Total = NDC + NAP,
         Pc_NDC = round((NDC / Total) * 100, 2),
         Pc_NAP = round((NAP / Total) * 100, 2))
```

## agriculture: n indicators and countries, by indicator type 
```{r}
indicators_ag <- indicators[indicators$SectorType == "Food, fiber, other ecosystem products", ]

ind_cty_type_ag <- indicators_ag %>%
  group_by(IndicatorType) %>%
  summarise(nIndicators = n(), 
            nCountries = n_distinct(Country), .groups = 'drop')

unique_cty <- unique(indicators_ag$Country)
print(unique_cty)

total_indicators <- 222
total_countries <- 10 

ind_cty_type_ag <- ind_cty_type_ag %>%
  mutate(pcIndicators = round((nIndicators / total_indicators) * 100),
         pcCountries = round((nCountries / total_countries) * 100))

ind_cty_type_ag <- ind_cty_type_ag[, c(1, 2, 4, 3, 5)]
```

## agriculture: n indicators by smart criteria and country
```{r}
#### convert from wide to long format
indicators_subset_ag <- indicators_ag[, c(2, 20:24)]
long_ind_ag <- indicators_subset_ag %>%
  pivot_longer(cols = -Country, names_to = "criteria", values_to = "value")

ind_smart_cty_ag <- long_ind_ag %>%
  group_by(criteria) %>%
  summarise(
    nIndicators = sum(value),
    nCountries = n_distinct(Country[value == 1])
  )

total_indicators <- 222
total_countries <- 10 

ind_smart_cty_ag <- ind_smart_cty_ag %>%
  mutate(pcIndicators = round((nIndicators / total_indicators) * 100),
         pcCountries = round((nCountries / total_countries) * 100))

ind_smart_cty_ag <- ind_smart_cty_ag[, c(1, 2, 4, 3, 5)]
```

### Save data to local files
Modify the file path to specify the location where you would like to store the new dataframe
```{r}
write.csv(ind_sector,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndicatorXsector.csv")
write.csv(ind_cty_type,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndic_nCtyXtype.csv")
write.csv(ind_funct_doc,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/1_nME_nCty_doc.csv")
write.csv(ind_smart_cty,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndicatorsXcriteriaXcty.csv")
write.csv(indicators_criteria_nr,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndicatorsXnCriteria.csv")
write.csv(indicators_smart,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndicatorsSMART.csv")
write.csv(ind_smart_doc,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nSMART_nCty_doc.csv")
write.csv(ind_cty_type_ag,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndic_nCtyXtype_Ag.csv")
write.csv(ind_smart_cty_ag,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_nIndicatorsXcriteriaXcty_ag.csv")
```


## Supplementary Table 3: indicators assessment for food and fiber sector
```{r}
indicators_ag$ISO3 <- countrycode(indicators_ag$Country, "country.name", "iso3c")
indicators_ag2 <- indicators_ag[, c(3, 4, 7, 19:24, 27)] 
indicators_ag2 <- indicators_ag2[, c(10, 1, 2, 3, 4, 5, 6, 7, 8, 9)]

# Save table to local file
# Modify the file path to specify the location where you would like to store the table
write.csv(indicators_ag2,"C:/Users/anowak/OneDrive - CGIAR/PhD/Ch1/Results/3_SI_Table3.csv")
```
